Implemented basic parsing for the request class
the parse_request function takes in a stream and parses it. The
return value is a hash where the first line of attributes and the
body of the request are first class members of the hash, while the
headers are in a nested hash.

Also, framework for the response and resource class have been
created. There is a basic work flow of the project in place, but
this is just an example.