Added a key 'query_string' to the  Request.parse_request hash
The value associated with 'query_string' is a Hash, and each
parameter of the query string portion of the request can be
accessed by its name (e.g. request_hash[:q] = > para for ?q=para)

Also, finished the testing of recieving a body as part of the
request. the :body key returns a string of the body